Move logic for rule to utils and change to suggest instead of fix
